requirements.txt
albumentations
fiftyone
fiftyone_brain
matplotlib
numpy
Pillow
pycocotools
torch
torchvision
tqdm
